backend/requirements.txt
fastapi==0.104.0
uvicorn==0.24.0
sqlmodel==0.0.14
sqlalchemy==2.0.23
psycopg[binary]==3.1.18
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.4.2
python-dotenv==1.0.0
pgvector==0.2.4
openai
